heavy work on optimization
added simulated annealing function, played around with several optimization functions, but all were either much worse or negligibly better than the original best-neighbor hill climb. however it seemed that the evaluation function was the root of the problem, not the optimization(as of now) so i implemented trigram analysis which seems to be working much better.